GAIN-3: Ignore test in/out files
I'll just have to force-add them if I'm going to source-control any.
I'm thinking that, when it comes time, the test framework can
dynamically download/expand the files if they're not already found
locally... but that's work for another day.  For now, I don't want to
see these test files in my status.

Changes to be committed:
	modified:   .gitignore

Changes not staged for commit:
	modified:   gallant_input/main.py

Untracked files:
	gallant_input/sigmfmetaparser.py