GAIN-3: Finished(?) initial SigMFMetaParser
- Added a "getter" to fetch the center frequency
- Wrote functionality to get a key from an object
- Removed DEBUGGING
- Added some private, common functionality
- Updated a docstring

GOOD NEWS: This appears to be working when tested with
manual input.

BAD NEWS: I just reread the JQR line item and I should be
answering questions about an IQ capture, *not* a SigMF
file capture.

Time to put a pin in this, merge, add backlog tickets,
and get after IQ file format parsing.

Changes to be committed:
	modified:   gallant_input/sigmfmetaparser.py

Changes not staged for commit:
	modified:   gallant_input/main.py