GAIN-5: Initial commit of the GAIN sub-package rds
It is here I will define all of my RDS-parsing code.
There is no guarantee any of this code works, much less runs.
I haven't even run the linters on it yet.

It remains to be determined if I will include this as a gallant_input
CLI argument or not.  For now, I intend it to be imported and utilized.
Hopefully, by a custom GNU Radio block.

At the very least, successful implementation of this design will
help me determine the success (or failure) of my RDS decode.
(And maybe even do it live with my GRC FM receiver...)

EOD commit

Changes to be committed:
	new file:   gallant_input/rds/__init__.py
	new file:   gallant_input/rds/block.py
	new file:   gallant_input/rds/block_id.py
	new file:   gallant_input/rds/constants.py